requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
google-auth-oauthlib==1.2.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
twilio==8.12.0
firebase-admin==6.3.0
razorpay==1.4.1
stripe==7.3.0
cryptography==41.0.7
requests==2.31.0
aiohttp==3.9.1
cors==1.0.1
python-multipart==0.0.6
